refactor of customEvent Action
