+++ packages/destination-actions/src/destinations/hubspot/customEvent/functions/hubspot-event-schema-functions.ts
@@ -1,6 +1,4 @@
-import { PayloadValidationError, IntegrationError, RetryableError, StatsContext } from '@segment/actions-core'
-import { SubscriptionMetadata } from '@segment/actions-core/destination-kit'
-import type { Payload } from './generated-types'
+import { PayloadValidationError, IntegrationError, RetryableError } from '@segment/actions-core'
 import {
   CreateEventDefinitionReq,
   CreatePropDefinitionReq,
@@ -11,102 +9,8 @@ import {
   CachableSchema,
   SchemaDiff,
   PropertyCreateResp
-} from './types'
-import { Client } from './client'
-import { LRUCache } from 'lru-cache'
-
-export const cache = new LRUCache<string, CachableSchema>({
-  max: 2000,
-  ttl: 1000 * 60 * 60
-})
-
-function stringFormat(str: string): StringFormat {
-  // Check for date or datetime, otherwise default to string
-  const isoDateTimeRegex =
-    /^(\d{4})-(0[1-9]|1[0-2])-(0[1-9]|[12]\d|3[01])(?:([ T])(\d{2}):?(\d{2})(?::?(\d{2})(?:[,\.](\d{1,}))?)?(?:(Z)|([+\-])(\d{2})(?::?(\d{2}))?)?)?$/ //eslint-disable-line no-useless-escape
-  const dateOnlyRegex = /^\d{4}-\d{2}-\d{2}$/ //eslint-disable-line no-useless-escape
-
-  if (isoDateTimeRegex.test(str)) {
-    return dateOnlyRegex.test(str) ? 'date' : 'datetime'
-  } else {
-    return 'string'
-  }
-}
-
-export function eventSchema(payload: Payload): Schema {
-  const { event_name, properties } = payload
-  const props: { [key: string]: SegmentProperty } = {}
-
-  if (properties) {
-    Object.entries(properties).forEach(([property, value]) => {
-      if (value !== null) {
-        props[property] = {
-          type: typeof value as SegmentPropertyType,
-          stringFormat: typeof value === 'string' ? stringFormat(value) : undefined
-        }
-      }
-    })
-  }
-  return { name: event_name, primaryObject: payload.record_details.object_type, properties: props }
-}
-
-export function getSchemaFromCache(
-  name: string,
-  subscriptionMetadata?: SubscriptionMetadata,
-  statsContext?: StatsContext
-): CachableSchema | undefined {
-  if (!subscriptionMetadata || !subscriptionMetadata?.actionConfigId) {
-    statsContext?.statsClient?.incr('cache.get.error', 1, statsContext?.tags)
-    return undefined
-  }
-
-  const schema: CachableSchema | undefined = cache.get(`${subscriptionMetadata.actionConfigId}-${name}`) ?? undefined
-  return schema
-}
-
-export async function saveSchemaToCache(
-  schema: CachableSchema,
-  subscriptionMetadata?: SubscriptionMetadata,
-  statsContext?: StatsContext
-) {
-  if (!subscriptionMetadata || !subscriptionMetadata?.actionConfigId) {
-    statsContext?.statsClient?.incr('cache.save.error', 1, statsContext?.tags)
-    return
-  }
-
-  cache.set(`${subscriptionMetadata.actionConfigId}-${schema.name}`, schema)
-  statsContext?.statsClient?.incr('cache.save.success', 1, statsContext?.tags)
-}
-
-export function compareSchemas(schema1: Schema, schema2: CachableSchema | undefined): SchemaDiff {
-  if (schema2 === undefined) {
-    return { match: 'no_match', missingProperties: {} }
-  }
-
-  if (schema1.name !== schema2.name && schema1.name !== schema2.fullyQualifiedName) {
-    throw new PayloadValidationError("Hubspot.CustomEvent.compareSchemas: Schema names don't match")
-  }
-
-  const missingProperties: { [key: string]: SegmentProperty } = {}
-
-  for (const [key, prop1] of Object.entries(schema1.properties)) {
-    const prop2 = schema2.properties[key]
-    if (prop2 === undefined) {
-      missingProperties[key] = prop1
-      continue
-    }
-    if (prop1.stringFormat === prop2.stringFormat && prop1.type === prop2.type) {
-      continue
-    } else {
-      throw new PayloadValidationError("Hubspot.CustomEvent.compareSchemas: Schema property types don't match")
-    }
-  }
-
-  return {
-    match: Object.keys(missingProperties).length > 0 ? 'properties_missing' : 'full_match',
-    missingProperties
-  }
-}
+} from '../types'
+import { Client } from '../client'
 
 export async function getSchemaFromHubspot(client: Client, schema: Schema): Promise<CachableSchema | undefined> {
   const response = await client.getEventDefinition(schema.name)
